% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genus.R
\docType{data}
\name{genus}
\alias{genus}
\title{Genus table.}
\format{Matrix with samples in columns and genera in rows.}
\source{
<https://www.datadryad.org/MECDRYAD>
}
\usage{
genus
}
\description{
Matrix of raw genus counts
}
\examples{
  genus
}
\keyword{datasets}
