% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{MEC sample information}
\format{Data frame with columns
\describe{
\item{id}{Sample ID.}
\item{age}{Age (years) at time of stool sample collection.}
\item{sex}{Sex.}
\item{ethnicity}{Self-reported ethnicity.}
\item{height}{Height (meters) at time of stool sample collection.}
\item{weight}{Weight (kilograms) at time of stool sample collection.}
\item{sample_month}{Month (1-12) of stool sample collection.}
\item{sample_year}{Year of stool sample collection.}
\item{sample_area}{Study site for stool sample collection.}
\item{batch}{Batch identifier for stool sample processing.}
}}
\source{
<https://www.datadryad.org/MECDRYAD>
}
\usage{
sample_data
}
\description{
Sample data for participants in the MEC microbiome study.
}
\examples{
  sample_data
}
\keyword{datasets}
