% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\docType{data}
\name{taxonomy}
\alias{taxonomy}
\title{Taxonomy names.}
\format{Data frame with columns
\describe{
\item{kingdom}{Kingdom name.}
\item{phylum}{Phylum name.}
\item{class}{Class name.}
\item{order}{Order name.}
\item{family}{Family name.}
\item{genus}{Genus name.}
\item{genus_id}{Unique genus identifier.}
}}
\source{
<https://www.arb-silva.de/>
}
\usage{
taxonomy
}
\description{
Names of taxa from kingdom to genus. Based on Silva database version ###
}
\examples{
  taxonomy
}
\keyword{datasets}
